dashboard/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
paramiko==3.4.0
httpx==0.26.0
python-dotenv==1.0.0
jinja2==3.1.2

